Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,9 +1,9 @@
 #include<stdio.h>
 int main()
 {
-    double point;
-    scanf("%f",&point);
-    printf("You entered: %f",point)
+    float point;
+    scanf("%le",&point);
+    printf("You entered: %le",point)
     return 0;
 
 }
